SpiderBots Main Program Launcher
The main program to launch the spiderbots program and instantiate all other classes. The agency is launched, and agents are created to match the robots. The agents are then added to the agency. The data commands are loaded and the audio controller and transcriber are enabled. The command parsing engine uses the transcription as input and command dictionary as reference to dictate actions to the agents. Agents also use object detectors to inform the server of targets and the environment variables such as colors, obstacles and distances.